+++ ErsatzTV.Core/Interfaces/Images/IImageCache.cs
@@ -5,7 +5,6 @@ namespace ErsatzTV.Core.Interfaces.Images;
 
 public interface IImageCache
 {
-    Task<Either<BaseError, byte[]>> ResizeImage(byte[] imageBuffer, int height);
     Task<Either<BaseError, string>> SaveArtworkToCache(Stream stream, ArtworkKind artworkKind);
     Task<Either<BaseError, string>> CopyArtworkToCache(string path, ArtworkKind artworkKind);
     string GetPathForImage(string fileName, ArtworkKind artworkKind, Option<int> maybeMaxHeight);